pyproject.toml
[project]
name = "rna-vel-pred"
version = "0.1.0"
description = "Add your description here"
readme = "README.rst"
license = { file = "LICENSE" }
authors = [
    {name = 'Taos Transue', email = 'taos.transue@utah.edu'},
    {name = 'Payton J. Thomas', email = 'p5thomas@ucsd.edu'},
]
requires-python = ">=3.10"
dependencies = [
    "h5py==3.8.0",
    "hydra-core>=1.3.2",
    "jupyterlab>=4.3.4",
    "lightning>=2.5.0.post0",
    "numpy<2",
    "pytest>=8.3.4",
    "scanpy==1.9.8",
    "scvelo==0.3.2",
    "seaborn>=0.13.2",
    "tables==3.8.0",
    "tensorboard>=2.18.0",
    "torch-geometric>=2.6.1",
    "torch>=2.5.1",
]
# uv pip install pyg_lib torch_scatter torch_sparse torch_cluster torch_spline_conv -f https://data.pyg.org/whl/torch-2.5.0+cu124.html

[project.scripts]
rna-vel-pred = "rna_vel_pred:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
